updated tests and api

+++ src/main/java/io/h2o/service/impl/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package io.h2o.service.impl;
 
 import io.h2o.domain.User;
+import io.h2o.exception.UserNotFoundException;
 import io.h2o.repository.UserRepository;
 import io.h2o.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -43,14 +45,30 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public boolean deleteUserSoft(Long id) {
-        int i = userRepository.deleteUserSoft(id);
-        if(i==0) return false;
-        else return true;
+//        Optional<User> user = userRepository.findById(id);
+//        if(!user.isPresent()) return false;
+//        else {
+//            userRepository.deleteUserSoft(id);
+//            return true;
+//        }
+        User user = userRepository.findById(id).orElse(null);
+        if(user!=null) return false;
+        else {
+            userRepository.deleteUserSoft(id);
+            return true;
+        }
+//        int i = userRepository.deleteUserSoft(id);
+//        if(i==0) return false;
+//        else return true;
     }
 
     @Override
-    public void deleteUserHard(Long id) {
-        User user = userRepository.findById(id).get();
-        userRepository.delete(user);
+    public boolean deleteUserHard(Long id){
+        User user = userRepository.findById(id).orElse(null);
+        if(user!=null) return false;
+        else {
+            userRepository.deleteUserSoft(id);
+            return true;
+        }
     }
 }
